refactor(core): Reduce code duplication in polling monitors
